+++ shyftDb/postgres_setup_test/drop_tables_test.psql
@@ -1,4 +1,5 @@
-DROP TABLE IF EXISTS internalTxs;
-DROP TABLE IF EXISTS txs;
-DROP TABLE IF EXISTS blocks;
-DROP TABLE IF EXISTS accounts;
+DROP TABLE IF EXISTS internalTxs CASCADE;
+DROP TABLE IF EXISTS txs CASCADE;
+DROP TABLE IF EXISTS blocks CASCADE;
+DROP TABLE IF EXISTS accounts CASCADE;
+DROP TABLE IF EXISTS accountblocks CASCADE;
